+++ Changelog.md
@@ -10,6 +10,7 @@
 - Support generation
   - Prune auto-generated supports with min distance constraint
   - Allow placing supports manually
+- Add anti-aliasing to slice preview
 
 ## v0.8.0 &mdash; June 30th, 2026